Commonize EX31 cached pipeline state
